+++ mycodo/mycodo_flask/app.py
@@ -68,7 +68,7 @@ def create_app(config=ProdConfig):
     def get_locale():
         misc = flaskutils.db_retrieve_table(app.config['MYCODO_DB_PATH'], Misc, first=True)
         if misc.language != '':
-            for key, value in LANGUAGES.iteritems():
+            for key, _ in LANGUAGES.iteritems():
                 if key == misc.language:
                     return key
         return request.accept_languages.best_match(LANGUAGES.keys())